DisplayQuiz code reorganization
