+++ src/app/NoteUseCaseFactory.java
@@ -1,10 +1,11 @@
 package app;
 
 import data_access.NoteDataAccessObject;
-import data_access.QuizDataAccessObject;
 import interface_adapter.ViewManagerModel;
 import interface_adapter.note.*;
-import interface_adapter.quiz.*;
+import interface_adapter.quiz.display.DisplayQuizzesController;
+import interface_adapter.quiz.display.DisplayQuizzesPresenter;
+import interface_adapter.quiz.display.DisplayQuizzesViewModel;
 import use_case.note.NoteDataAccessInterface;
 import use_case.note.delete.DeleteNoteInputBoundary;
 import use_case.note.delete.DeleteNoteInteractor;
@@ -13,11 +14,10 @@ import use_case.note.save.SaveNoteInputBoundary;
 import use_case.note.save.SaveNoteInteractor;
 import use_case.note.save.SaveNoteOutputBoundary;
 
-import use_case.quiz.DisplayQuizzesInputBoundary;
-import use_case.quiz.DisplayQuizzesInteractor;
-import use_case.quiz.DisplayQuizzesOutputBoundary;
+import use_case.quiz.display.DisplayQuizzesInputBoundary;
+import use_case.quiz.display.DisplayQuizzesInteractor;
+import use_case.quiz.display.DisplayQuizzesOutputBoundary;
 import use_case.quiz.QuizDataAccessInterface;
-import view.DisplayQuizzesView;
 import view.NoteView;
 
 public class NoteUseCaseFactory {
