+++ src/interface_adapter/quiz/display/DisplayQuizzesViewModel.java
@@ -1,4 +1,4 @@
-package interface_adapter.quiz;
+package interface_adapter.quiz.display;
 import interface_adapter.ViewModel;
 
 import java.beans.PropertyChangeListener;
